Noise_Texture/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noise-texture"
version = "1.0.0"
description = "Comprehensive noise texture generation library for computer graphics"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "RoboCute Project"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["noise", "texture", "procedural", "graphics", "perlin", "simplex", "worley"]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "noise>=1.2.1",
    "opensimplex>=0.4.5",
    "scipy>=1.7.0",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.5.0"]
perf = ["numba>=0.56.0"]
dev = ["pytest>=7.0.0", "black>=22.0", "flake8>=4.0"]

[project.urls]
Homepage = "https://github.com/robocute/noise-texture"
Documentation = "https://noise-texture.readthedocs.io"

[tool.setuptools]
packages = ["noise_texture"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
